requirements.txt
aiogram
aiohttp
aiosqlite
asynclog
beautifulsoup4
python-dateutil
python-dotenv
requests
